backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
idna==3.3
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
